Fix empty STDIN bug
After a lot of investigation, it turns out that instantiating CGI wipes
out STDIN. So any plugin that defines a POST/PUT endpoint with a request
body will receive and empty body that fails validation.

The solution is to realise that a plugin class inherits an already
instantiated CGI object from it's base Koha::Plugins::Base class, which
it can access via $self->{cgi}, removing the need to instantiate a new
one.